Function k7lib::string::next_permutation

pub fn next_permutation<T>(s: &mut [T]) -> bool where
    T: Ord, 

Permutes the slice into the next greater permutation in lexicographically order.

If successful, true is returned. If the next permutation does not exist, nothing is performed and false is returned.

Examples

let mut s = [0, 0, 2, 1];
assert!(next_permutation(&mut s));
assert_eq!(s, [0, 1, 0, 2]);

let mut s = [2, 1, 0, 0];
assert!(!next_permutation(&mut s));
assert_eq!(s, [2, 1, 0, 0]);

Enumerating all permutations:

let mut s = [0, 0, 1, 2]; // sorted
loop {
  println!("{:?}", s);
  if !next_permutation(&mut s) {
    break;
  }
}

Time complexity

$\Theta(n)$ worst per operation

$\Theta(n!)$ worst, to enumerate all permutations